[Blazor] Try to load the assembly when finding out a root component or a parameter in webassembly (#52331) (#52351)
# Load assembly when finding root component or parameter in WebAssembly

This change allows the system to load the assembly when a root component is defined in an RCL and rendered in WebAssembly, even if the assembly is not yet loaded into memory.

## Description

When a root component is defined in an RCL and rendered in WebAssembly, it might happen that the assembly for the component is not yet loaded into memory if the app has not used any type from the dll.
This causes the framework to fail finding the assembly when it tries to look for it in the list of loaded assemblies.

The fix is to detect this in webassembly and try to load the root component type at that point. Same with the component parameters.

Fixes #52129

## Customer Impact

A Blazor Web app may experience a failure trying to render a root component in WebAssembly, when the component is defined in a Razor Class Library. This will happen if the assembly containing the component hasn't yet been loaded on the client, and the app is trying to render that component.

## Regression?

- [ ] Yes
- [x] No

## Risk

- [ ] High
- [ ] Medium
- [x] Low

The assembly will already be downloaded by JS and is ready to be loaded.

## Verification

- [x] Manual (required)
- [x] Automated

There was manual verification that the fix addressed the issue and it was automated into an E2E test.

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A

There are no packaging changes associated with this fix.

## When servicing release/2.1

- [ ] Make necessary changes in eng/PatchConfig.props