[Infrastructure] Move back to NPM (#52914)
This change removes Yarn from our infrastructure completely and replaces it with NPM. There is a package.json in the top folder of the respository that defines the workspace. The workspace lists all the paths to packages that are part of the repository. In addition to that, the top package.json defines scripts that can be run from the root of the repository to install, build test, etc. all the packages in the repository.

Each package has its own package.json file that defines its dependencies and scripts. The scripts in the top package.json file are just wrappers around the scripts in the individual package.json files.

## Developer workflow

### Restore dependencies

To install the dependencies, you can run npm ci from the root of the repository. This will install all the dependencies for all the packages in the repository and automatically link them together. This normally takes very little time once you've have downloaded the packages onto your machine (10s for me).

### Install or update new dependencies
Run npm install ... on whatever node you want to install the dependencies. The package-lock.json will get updated accordingly. For installing new dependencies you might need to authenticate yourself with the NPM VSTS feed. To do that, run:

* `npm install -g vsts-npm-auth`
* `vsts-npm-auth -config .npmrc -F`. This will walk you through the auth process and provision the required credentials.
*  `npm install`

### Updating deps to fix Component Governance alerts

* Run npm audit to see the list of issues.
* Run npm audit fix to see if the issues can get addressed automatically.
* If that's not the case:
  * If possible, update to a newer version of the package that doesn't include the vulnerable dependency.
  * If that's not possible, at the appropriate node, add an entry to the "overrides" property to force the transitive dependency into a non-vulnerable version.
  * run npm install from the root.
  * run npm audit again to see if the issues have disappeared, if they haven't continue making changes (update more packages, etc).

Upgrading the main deps should always be the preferred route over using overrides. Overrides require that we periodically remove them to check if we can update the dependency later on.

### Building JS components

You can run `npm run build` at the repo root and it will build all the projects (takes about 2 minutes). Or alternatively, you can run `npm run build` on an individual project.json to build only that `project.json`. If you do this, any dependent project should have been built previously. Dependencies won't get build automatically.

### Testing JS components

There are two test categories:
 * Unit tests: They are self-contained and don't depend on any of the .NET Code. Run `npm run test`.
 * Integration tests: They test the JS in combination with the .NET Code, for which you first need to have run `eng\Build.cmd`, before you invoke the tests. To execute them, run `npm run integration-test`