Stop duplicating history with streaming-enhanced-form PRG. Fixes #51674 (#53301)
# Stop duplicating history with streaming-enhanced-form PRG

Makes the post-redirect-get pattern work without duplicating history entries.

## Description

The PRG pattern is common but has the annoying disbenefit that, after submitting a form, the browser's "back" button doesn't work as normal because there is a duplicated history entry. The user has to click "back" twice or more times depending on how many times they have submitted the form.

In the case of enhanced navigation, Blazor fixes this by explicitly avoiding duplicated history entries. But this wasn't implemented for the streaming+enhanced case. This PR adds the avoid-duplication behavior in that case too.

Fixes #51674

## Customer Impact

Without this, the PRG pattern would work nicely on an enhanced form, but if you then also enabled streaming SSR for that form, it would go back to duplicating history entries and forcing the user to click "back" multiple times.

**NOTE**: This PR still doesn't implement non-duplication in the non-enhanced-nav case (with or without streaming). That's out of scope because without enhanced nav, our code doesn't control how the response is handled, so we have to accept the default browser behavior (which does duplicate the history entries after PRG).

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [x] Medium
- [ ] Low

There's some pretty intricate logic for the interactions between streaming, enhanced nav, redirections, and the history stack. This adds another case. Mitigations:

 * We have pretty extensive tests for all the combinations we can think of
 * The way the change is structured, it *should* not impact other cases

## Verification

- [x] Manual (required)
- [x] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A