Add title to column options. Fixes #52141 (#53299)
# Add title to column options

In QuickGrid, adds a title attribute to the column options button. Required for accessibility.

## Description

Attaches the title "Column options" to the column options button. Fixes the "Buttons must have discernible text" accessibility requirement.

Fixes #52141

## Customer Impact

Previously the button only showed an icon, which is not helpful for screen readers. After this PR you can hover to see the tooltip, and it may be read by assistive technologies.

## Regression?

- [ ] Yes
- [x] No

[If yes, specify the version the behavior has regressed from]

## Risk

- [ ] High
- [ ] Medium
- [x] Low

It's literally just adding an HTML attribute that has no other effect.

## Verification

- [x] Manual (required)
- [ ] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [x] N/A