[Blazor] Dispatch inbound activity handlers in Blazor Web apps (#53185) (#53313)
# Dispatch inbound activity handlers in Blazor Web apps

Fixes an issue where inbound activity handlers were not getting invoked in Blazor Web apps.

## Description

Inbound activity handlers were a new feature introduced in .NET 8 to allow Blazor Server apps to monitor circuit activity and make services available to Blazor components that wouldn't otherwise be accessible. However, we also introduced a new "Blazor Web" app model that didn't incorporate this new feature correctly.

This PR fixes the issue by:
* Building the inbound activity pipeline in Blazor Web apps just after circuit handlers are retrieved
* Invoking inbound activity handlers in web root component updates

Fixes #51934

## Customer Impact

Multiple customers have indicated being affected by this issue. Since this is a new feature in .NET 8, customers expect it to work in Blazor Web apps, which we recommend for new projects. Without this fix, the functionality only works in traditional Blazor Server apps.

## Regression?

- [ ] Yes
- [X] No

This is a new feature in .NET 8.

## Risk

- [ ] High
- [ ] Medium
- [X] Low

The fix is simple and applies to a well-tested area of the framework, so our E2E tests are likely to catch issues with this change.

## Verification

- [X] Manual (required)
- [X] Automated

## Packaging changes reviewed?

- [ ] Yes
- [ ] No
- [X] N/A