Introduce a read-only mode for data protection keyring consumers (#53539)
* Introduce a read-only mode for data protection keyring consumers

When multiple app instances consume the same keyring, they all try to rotate it, leading to races.  This change introduces an IConfiguration property (usually set as an env var) that puts data protection in a read-only mode.  The expectation is that writing will be done by a separate (i.e. non-app-instance) component.

Fixes #52915