[Blazor] Enable regex constraint in Blazor routing (#53533)
This was discovered when working on the issue for #53138. The regex
constraint is not enabled by default on the blazor router.

As a result if a route uses a regex constraint, it will work on SSR but
will fail the moment the Blazor router tries to construct the route.

The fix enables the regex constraint on the blazor router,
unconditionally for server, and behind a feature flag for webassembly.

This is because enabling the regex constraint adds +80kb to the payload
due to the inclusion of the System.Text.RegularExpressions assembly.