[release/8.0] [Blazor] Fix hot reload memory leak (#53928)
* Resolve change token leak in Blazor hot reload (#53750)

Fix of razor hotreload change token leak. This disposes the
old change tokens after the ClearCache event or before overwriting. 
If something goes wrong and this isn't cleared
before the next invocation of UpdateEndpoints on the
razor data source, clear it and dispose of it then.

* Add unit test to confirm change token is disposed during (#53827)

* Add unit test to confirm change token is disposed during
razer hot reload.

* Per Makinnon's feedback, switch to a callback model
to create the wrapped disposable for this unit test.

* Update src/Components/Endpoints/test/HotReloadServiceTests.cs

---------

Co-authored-by: Mackinnon Buck <mackinnon.buck@gmail.com>

---------

Co-authored-by: jacdavis <jacdavis@users.noreply.github.com>