Enable portable cross-arch source-building to unknown architectures.
The repository supports building for unknown rids to enable building for
non-portable rids. This build configuration is under the
'$(PortableBuild)' == 'false' condition.

This changes those conditions to '$(DotNetBuildSourceOnly)' == 'true'
to enable building a portable configuration for an unknown portable rid.
This configuration is used when cross-building for an architecture
that is not yet known to the SDK, like linux-ppc64le or linux-s390x.