Reuse known keys in XmlKeyManager
Keys read from the IXmlRepository (as opposed to created) use deferred decryption but may still be decrypted multiple times.  Since they're immutable (other than revocation), keep track of the keys we've seen and reuse them next time they're returned from the IXmlRepository.

In the absence of deletion support, we don't expect the list of keys to ever shrink, so there's no reason to prune the cache.

While this is likely a perf win, it's a small perf win and only once a day, so it would probably not be worthwhile on that basis alone.  However, this should also reduce the number of calls to `IXmlDecryptor.Decrypt` (which may, e.g., make an Azure KeyVault request).  Each time the key ring is refreshed, it requests all keys from the key manager, clobbering all keys that may have already been decrypted (at least the default key will have been).  In the single app-instance case, for example, there should be no need to call Decrypt at all for any key created during the current execution of the app (old keys will have to be decrypted the first time they're used).

This shouldn't affect security because (a) the key ring already stores all the `IKeys` all the time and (b) `IKey` wraps the actual data in an `ISecret`.

It would have been nice to pass the key ring's list of known keys to the key manager, rather than having it store its own cache, but the key ring's storage format would require abstraction violations that would work poorly with our public extension points.