Limit parallelism of forced key ring refreshes
During the auto-refresh window, an `Unprotect` call with an unknown key ID is allowed to bypass the cache and perform a fresh read of the key ring (#11987).  While, in principle, "force" means that it must happen every time, in practice, it doesn't make sense to block until an ongoing refresh is finished and then immediately perform another one.  Instead, wait for the in-progress refresh to complete and return the same value (or throw, as appropriate).

Preserving the precise exception from the thread that actually does the refresh didn't seem important, so I didn't bother (for now).