Enable building native projects for windows arm64 (#55145)
Fixes building aspnetcore for windows arm64 in the VMR.

## Description

Right now the native projects are only built when targetting x86/x64 architectures. This causes an issue in the VMR because we're building windows arm64 in a separate job and aspnetcore fails because of the missing native assets.

The desktop msbuild build for the native projects was disabled for arm64, enable it to match the other architectures.