Do not disable JWT validation unless requested
Prior to this fix, when JwtBearerOptions are read from the default
config section, issuer and audience validation would be disabled if no
valid issuers or valid audiences were explicitly configured from config.

This was a problem if the authority was supplied from config because
this would disable validating the issuer matched what was provided by
the authority. In the case of Entra, this means tokens for another
tenant could be used since the JWKS are the same across tenants.

Disabling audience validation when no valid audiences are provided by
config is less severe because valid audiences cannot be inferred from an
authority, but it could still be supplied outside config, and it should
still validated unless validation is explicitly opted out of.

The configuration reading logic now reads "ValidateIssuer" and
"ValidateAudience" booleans from config allowing developers to
explicitly opt out of validation by setting these false. Otherwise,
the default remains true.