Add a key management simulator for data protection
It's not properly a sample, since it requires IVT, but it's much easier to read and use as a standalone app than as a test.  (It would also make a poor test, since perfect performance is not expected and thus there's no way to define success.)

Simulates running an app with a given number of instances for a given number of days and computes a score (lower is better) indicating how likely missing key errors were during that simulated time period.  Also tracks calls to the storage and encryption backends so we can see whether e.g. retries have a significant perf impact.

In the future, it would be nice if instances started and stopped over the course of the run, rather than all living for the entire duration.