Annotate SignalR server for native AOT
Fix SignalR Server's usage of MakeGenericMethod when using a streaming reader (IAsyncEnumerable or ChannelReader) following the same approach as the client. Add a runtime check and throw an exception when trying to stream a ValueType in native AOT.

Adjust the public annotations:
* Remove RequiresUnreferencedCode from AddSignalR
* Add RequiresUnreferencedCode to MessagePack and NewtonsoftJson protocols

Support trimming and AOT in DefaultHubDispatcher by adding a feature switch to turn off custom awaitable support.
Update ObjectMethodExecutor to support trimming and AOT by a new method that doesn't look for custom awaitables, and uses reflection instead of Linq.Expressions.