Simply SignalR Client IAsyncEnumerable usage
Today we create a separate "cancelable" async enumerable so we can "link" the passed in CancellationTokenSource with the CancellationToken passed into GetAsyncEnumerator. We don't need to do this in this scenario because we are the ones enumerating over the object, and we aren't passing in a CancellationToken during the enumeration.

Instead, we can simply enumerate over the supplied IAsyncEnumerable and call WithCancellation, passing in the supplied CancellationTokenSource.